requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
aiosqlite>=0.20.0
bcrypt>=4.1.0
PyJWT>=2.8.0
slowapi>=0.1.9
